Created the Chess\Piece\Capablanca namespace
