+++ src/Piece/Capablanca/A.php
@@ -1,25 +1,23 @@
 <?php
 
-namespace Chess\Piece;
+namespace Chess\Piece\Capablanca;
 
 use Chess\Piece\AbstractPiece;
-use Chess\Piece\CapablancaTrait;
+use Chess\Piece\B;
 use Chess\Piece\N;
-use Chess\Piece\R;
-use Chess\Piece\RType;
 use Chess\Variant\Capablanca\PGN\AN\Piece;
 use Chess\Variant\Capablanca\PGN\AN\Square;
 
-class C extends AbstractPiece
+class A extends AbstractPiece
 {
     use CapablancaTrait;
 
     public function __construct(string $color, string $sq, Square $square)
     {
-        parent::__construct($color, $sq, $square, Piece::C);
+        parent::__construct($color, $sq, $square, Piece::A);
 
         $this->mobility = [
-            ...(new R($color, $sq, $square, RType::SLIDER))->mobility,
+            ...(new B($color, $sq, $square))->mobility,
             'knight' => (new N($color, $sq, $square))->mobility,
         ];
     }
