+++ src/Piece/Capablanca/C.php
@@ -1,24 +1,24 @@
 <?php
 
-namespace Chess\Piece;
+namespace Chess\Piece\Capablanca;
 
 use Chess\Piece\AbstractPiece;
-use Chess\Piece\CapablancaTrait;
-use Chess\Piece\B;
 use Chess\Piece\N;
+use Chess\Piece\R;
+use Chess\Piece\RType;
 use Chess\Variant\Capablanca\PGN\AN\Piece;
 use Chess\Variant\Capablanca\PGN\AN\Square;
 
-class A extends AbstractPiece
+class C extends AbstractPiece
 {
     use CapablancaTrait;
 
     public function __construct(string $color, string $sq, Square $square)
     {
-        parent::__construct($color, $sq, $square, Piece::A);
+        parent::__construct($color, $sq, $square, Piece::C);
 
         $this->mobility = [
-            ...(new B($color, $sq, $square))->mobility,
+            ...(new R($color, $sq, $square, RType::SLIDER))->mobility,
             'knight' => (new N($color, $sq, $square))->mobility,
         ];
     }
