+++ tests/unit/Piece/Capablanca/CTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Chess\Tests\Unit\Piece;
+namespace Chess\Tests\Unit\Piece\Capablanca;
 
-use Chess\Piece\C;
+use Chess\Piece\Capablanca\C;
 use Chess\Tests\AbstractUnitTestCase;
 use Chess\Variant\Capablanca\PGN\AN\Square;
